pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "messenger-logger-callback"
version = "0.2.0"
authors = [
  { name="Riko0", email="grigoriyalexeenko@gmail.com" },
]
description = "A training logger with Hugging Face Trainer integration and standalone API for sending logs to a remote server."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "requests>=2.25.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
trainer = ["transformers>=4.0.0"]

[project.urls]
Homepage = "https://github.com/Riko0/messenger-logger-callback"
"Bug Tracker" = "https://github.com/Riko0/messenger-logger-callback/issues"
